+++ src/index.js
@@ -1,7 +1,8 @@
 import React from "react";
-import { render } from "react-dom";
-import Switch from "app/Switch";// eslint-disable-line import/no-unresolved
+import {render} from "react-dom";
+import Switch from "app/Switch";
 import Application from "robe-react-commons/lib/application/Application";
+import Locale from "assets/en_US.json";// eslint-disable-line import/no-unresolved
 
 const app = document.getElementById("app");
 
@@ -13,6 +14,7 @@ const app = document.getElementById("app");
 
 
 Application.setBaseUrlPath("http://127.0.0.1:8081/robe");
+Application.loadI18n(Locale);
 
 
 render((<Switch />), app);